backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM & AI
anthropic==0.7.7
openai==1.3.7
langchain==0.1.0
langchain-anthropic==0.1.0
tiktoken==0.5.2

# Document processing
pypdf==3.17.1
python-docx==1.1.0
python-magic==0.4.27

# PDF generation
weasyprint==60.1
jinja2==3.1.2

# Storage
boto3==1.34.7
minio==7.2.0

# Utilities
httpx==0.25.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
